Fix an obsolete dependency libsystemd-daemon
The libsystemd-daemon component referred to from within configure.ac is not available in Ubuntu Trusty which is referenced as a recommended platform for building the project.
The replacement component is libsystemd.
I'm not completely sure that the required minimal version number of 37 is still relevant, the libsystemd version on my Trusty machine reported by pkg-config is 229 which is way higher that the mentioned 37. Build successfully completes after the change. The 'check' make target does not, but that's a different story.

https://at.projects.genivi.org/jira/browse/NSM-1